+++ test/data/languages/scala/program.scala
@@ -1,0 +1,3 @@
+object program extends App {
+  println("hello from scala!")
+}