+++ doc/about/changes.md
@@ -7,7 +7,7 @@ in progress
 ### New features
 
 - Add `header` argument to `system_package()` to find header files
-- Support Java
+- Support Java/Scala
 
 ---