+++ doc/about/changes.md
@@ -8,6 +8,7 @@ in progress
 
 - Add `header` argument to `system_package()` to find header files
 - Support Java/Scala
+- Add support for user-defined arguments
 
 ---